+++ mudacl/src/main/java/com/google/daq/orchestrator/mudacl/AclProvider.java
@@ -2,11 +2,15 @@ package com.google.daq.orchestrator.mudacl;
 
 import com.google.daq.orchestrator.mudacl.DeviceTopology.MacIdentifier;
 import com.google.daq.orchestrator.mudacl.DeviceTypes.DeviceClassifier;
+import com.google.daq.orchestrator.mudacl.SwitchTopology.Ace;
 import com.google.daq.orchestrator.mudacl.SwitchTopology.Acl;
+import java.util.Collection;
 import java.util.List;
 
 public interface AclProvider {
-  Acl makeEdgeAcl(MacIdentifier macAddress, DeviceClassifier device);
+  Acl makeEdgeAcl(DeviceClassifier device, MacIdentifier macAddress);
+
+  Acl makeUpstreamAcl(DeviceClassifier device, MacIdentifier macAddress);
 
   List<String> targetTypes();
 }