+++ mudacl/src/main/java/com/google/daq/orchestrator/mudacl/DeviceTopology.java
@@ -5,8 +5,10 @@ import java.util.TreeMap;
 
 public class DeviceTopology {
 
+  private static final String UPSTREAM_FORMAT = "dp_%s_upstream";
   private static final String DP_PORT_FORMAT = "dp_%s_port_%d";
-  private static final String TEMPLATE_FORMAT = "template_%s";
+  private static final String EDGE_TEMPLATE_FORMAT = "template_%s";
+  private static final String UPSTREAM_TEMPLATE_FORMAT = "upstream_%s";
 
   public Map<MacIdentifier, Placement> macAddrs = new TreeMap<>();
 
@@ -24,14 +26,20 @@ public class DeviceTopology {
     public String dpName;
     public Integer portNum;
 
-    public String toString() {
+    boolean isTemplate() {
+      return portNum == null;
+    }
+
+    public String edgeName() {
       return isTemplate()
-          ? String.format(TEMPLATE_FORMAT, dpName)
+          ? String.format(EDGE_TEMPLATE_FORMAT, dpName)
           : String.format(DP_PORT_FORMAT, dpName, portNum);
     }
 
-    boolean isTemplate() {
-      return portNum == null;
+    public String upstreamName() {
+      return isTemplate()
+          ? String.format(UPSTREAM_TEMPLATE_FORMAT, dpName)
+          : String.format(UPSTREAM_FORMAT, dpName, portNum);
     }
   }
 }