+++ docs/changelog.md
@@ -8,8 +8,10 @@
 	* Move more configuration to faucet.yaml files.
 	* Enable device groups.
 	* Clean up testing infrastructure.
-* 0.9.0
+* 0.8.2
 	* Update to use python3 (may require bin/clean for local setups).
 	* Adding docs for test lab setup.
 	* Improvements to docker test stability.
 	* Purge old docker containers after new image download.
+	* MUD file generator error message improvements.
+	* Add -n (no_test) mode to just put devices on network.